+++ proy_client/src/pages/admin/Usuarios/index.js
@@ -1,0 +1,1 @@
+export {default} from "./Usuarios";
